+++ src/cell/build.c
@@ -907,7 +907,6 @@ write_manifests(build_t* build)
 			visor_warn(build->visor, "'apiLevel': missing value: targeting API level 1");
 	}
 
-	// note: SGMv1 encodes the resolution width and height as separate fields.
 	jsal_get_prop_string(-6, "resolution");
 	if (!jsal_is_string(-1)
 		|| sscanf(jsal_to_string(-1), "%dx%d", &width, &height) != 2)
@@ -962,10 +961,11 @@ write_manifests(build_t* build)
 		visor_warn(build->visor, "'saveID': no save ID - '~/' prefix will be disabled");
 	}
 
-	// write game.sgm (legacy manifest, for compatibility with Sphere 1.x)
-	// note: SGM requires the main script path to be relative to '@/scripts'.
-	//       this differs from Sphere v2 (game.json), where it's relative to '@/'.
+	// write game.sgm (main Sphere game manifest)
+	// note: `script` (Sphere 1.x-compatible main script) is relative to '@/scripts'.
+	//       this differs from Sphere v2 `main`, which is relative to '@/'.
 	file = fs_fopen(build->fs, "@/game.sgm", "wb");
+	fprintf(file, "version=%d\n", api_version);
 	fprintf(file, "name=%s\n", jsal_to_string(-8));
 	fprintf(file, "author=%s\n", jsal_to_string(-7));
 	fprintf(file, "description=%s\n", jsal_to_string(-6));
@@ -976,10 +976,15 @@ write_manifests(build_t* build)
 		fprintf(file, "script=%s\n", path_cstr(script_path));
 		path_free(script_path);
 	}
+	else {
+		script_path = fs_relative_path(path_cstr(main_path), "@/");
+		fprintf(file, "resolution=%dx%d\n", width, height);
+		fprintf(file, "main=%s\n", path_cstr(script_path));
+	}
 	fclose(file);
 	jsal_pop(8);
 
-	// write game.json (Sphere v2 JSON manifest)
+	// write game.json (Sphere v2 JSON metadata)
 	jsal_stringify(-1);
 	json_text = jsal_get_lstring(-1, &json_size);
 	fs_fspew(build->fs, "@/game.json", json_text, json_size);